出身校情報登録機能の修正

+++ app/controllers/public/went_schools_controller.rb
@@ -11,6 +11,7 @@ class Public::WentSchoolsController < ApplicationController
       flash[:notice] = "出身校情報を登録しました"
       redirect_to user_path(current_user)
     else
+      flash[:alert] = went_school.errors.full_messages.join(", ")
       render :new
     end
   end
@@ -22,7 +23,10 @@ class Public::WentSchoolsController < ApplicationController
   def update
     went_school = current_user.went_school
     if went_school.update(went_school_params)
-      flash[:notice] = "出身校情報を登録しました"
+      flash[:notice] = "出身校情報を更新しました"
+      if current_user.review.present?
+        current_user.review.destroy
+      end
       redirect_to user_path(current_user)
     else
       @went_school = current_user.went_school
@@ -33,7 +37,10 @@ class Public::WentSchoolsController < ApplicationController
   def destroy
     went_school = current_user.went_school
     went_school.destroy
-    current_user.review.destroy
+    if current_user.review.present?
+        current_user.review.destroy
+    end
+    flash[:notice] = "出身校情報を削除しました"
     redirect_to user_path(current_user)
   end
